Implement functionality for inputting amount for credit card payments.
