+++ src/styles/pages/_user-view.scss
@@ -98,7 +98,7 @@ nav {
 
   .button--small {
     padding: 10px 20px;
-    font-size: 0.875rem;
+    font-size: 2.5rem;
   }
 
   .button--green {
